Expanding on roles and decision making
See also #1. This PR mostly aims to document what the current roles and
processes for decision making and getting things merged are, so that
they will be clearer to the wider community. That is, most of what I
have written here is intended to describe the current model (rather than
prescribe a new model).

Unlike in #1, I haven't described a "Language Supporter" role: in order
to keep this PR small, I would prefer to focus just on people who have
commit access to the compiler and/or core libraries for now.

There is also a part of this PR which does (sort of) represent a change
in the governance model, however. I would also like to start giving
commit access (i.e. "core collaborator" status) to some people on the
compiler repository. Up until now, we haven't done this at all; everyone
who has commit access on the compiler repository is currently a "core
owner".

I would like to add more core collaborators because we are a little
overstretched: we are not recruiting new maintainers at a very high
rate, and I think there is a bit too much going on for us (i.e. the
current core team) to be able to comfortably keep up with right now.

I think the main barrier to offering more people commit access on the
compiler repo right now is that it is not entirely clear what would be
expected of them. With the changes in this PR, I hope to make
expectations clear, so that we can start adding people as core
collaborators for the compiler repository, and so that these people will
feel empowered to help out without worrying about overstepping
expectations.

Finally, there are also some changes to make the formatting of this
document more consistent, and some minor clarifications (e.g. in the
Purpose section).